:bug: Fix selection on new/rename note
Chokidar does not support rename event, and implementing that oneself
is far from trivial having to deal with timing issues/assumption (even
when using baconjs).

Instead, assume that new notes to not be created randomly while the
user is active in Atom, so a new file means new note -or- rename (i.e.
unlink => add).